+++ eng/packages/http-client-csharp/generator/Azure.Generator/src/Primitives/NewAzureProjectScaffolding.cs
@@ -7,6 +7,7 @@ using Microsoft.TypeSpec.Generator.Primitives;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 
 namespace Azure.Generator.Primitives
 {
@@ -18,6 +19,7 @@ namespace Azure.Generator.Primitives
         private const string MSBuildThisFileDirectory = "$(MSBuildThisFileDirectory)";
         private const string RelativeCoreSegment = "sdk/core/Azure.Core/src/Shared/";
         private const string ParentDirectory = "../";
+        private const string SharedSourceLinkBase = "Shared/Core";
 
         /// <inheritdoc/>
         protected override string GetSourceProjectFileContent()
@@ -39,35 +41,52 @@ namespace Azure.Generator.Primitives
             int pathSegmentCount = GetPathSegmentCount();
             if (AzureClientGenerator.Instance.InputLibrary.InputNamespace.Auth?.ApiKey is not null)
             {
-                builder.CompileIncludes.Add(new CSharpProjectWriter.CSProjCompileInclude(GetCompileInclude("AzureKeyCredentialPolicy.cs", pathSegmentCount), "Shared/Core"));
+                builder.CompileIncludes.Add(new CSharpProjectWriter.CSProjCompileInclude(GetCompileInclude("AzureKeyCredentialPolicy.cs", pathSegmentCount), SharedSourceLinkBase));
             }
 
-            TraverseInput(out bool hasOperation, out bool hasLongRunningOperation);
+            bool hasOperation = false;
+            bool hasLongRunningOperation = false;
+            foreach (var client in AzureClientGenerator.Instance.InputLibrary.InputNamespace.Clients)
+            {
+                TraverseInput(client, ref hasOperation, ref hasLongRunningOperation);
+            }
+
             if (hasOperation)
             {
-                builder.CompileIncludes.Add(new CSharpProjectWriter.CSProjCompileInclude(GetCompileInclude("RawRequestUriBuilder.cs", pathSegmentCount), "Shared/Core"));
+                foreach (var file in _operationSharedFiles)
+                {
+                    builder.CompileIncludes.Add(new CSharpProjectWriter.CSProjCompileInclude(GetCompileInclude(file, pathSegmentCount), SharedSourceLinkBase));
+                }
             }
 
             if (hasLongRunningOperation)
             {
                 foreach (var file in _lroSharedFiles)
                 {
-                    builder.CompileIncludes.Add(new CSharpProjectWriter.CSProjCompileInclude(GetCompileInclude(file, pathSegmentCount), "Shared/Core"));
+                    builder.CompileIncludes.Add(new CSharpProjectWriter.CSProjCompileInclude(GetCompileInclude(file, pathSegmentCount), SharedSourceLinkBase));
                 }
             }
 
             return builder.Write();
         }
 
-        private static readonly IReadOnlyList<string> _lroSharedFiles =
+        private static readonly IReadOnlyList<string> _operationSharedFiles =
         [
+            "RawRequestUriBuilder.cs",
+            "TypeFormatters.cs",
+            "RequestHeaderExtensions.cs",
             "AppContextSwitchHelper.cs",
-            "AsyncLockWithValue.cs",
-            "FixedDelayWithNoJitterStrategy.cs",
             "ClientDiagnostics.cs",
             "DiagnosticScopeFactory.cs",
             "DiagnosticScope.cs",
             "HttpMessageSanitizer.cs",
+            "TrimmingAttribute.cs",
+        ];
+
+        private static readonly IReadOnlyList<string> _lroSharedFiles =
+        [
+            "AsyncLockWithValue.cs",
+            "FixedDelayWithNoJitterStrategy.cs",
             "IOperationSource.cs",
             "NextLinkOperationImplementation.cs",
             "OperationFinalStateVia.cs",
@@ -77,25 +96,25 @@ namespace Azure.Generator.Primitives
             "OperationPoller.cs",
             "SequentialDelayStrategy.cs",
             "TaskExtensions.cs",
-            "TrimmingAttribute.cs",
             "VoidValue.cs"
         ];
 
-        private static void TraverseInput(out bool hasOperation, out bool hasLongRunningOperation)
+        private static void TraverseInput(InputClient rootClient, ref bool hasOperation, ref bool hasLongRunningOperation)
         {
             hasOperation = false;
             hasLongRunningOperation = false;
-            foreach (var inputClient in AzureClientGenerator.Instance.InputLibrary.InputNamespace.Clients)
+            foreach (var method in rootClient.Methods)
             {
-                foreach (var method in inputClient.Methods)
+                hasOperation = true;
+                if (method is InputLongRunningServiceMethod || method is InputLongRunningPagingServiceMethod)
                 {
-                    hasOperation = true;
-                    if (method is InputLongRunningServiceMethod || method is InputLongRunningPagingServiceMethod)
-                    {
-                        hasLongRunningOperation = true;
-                        return;
-                    }
+                    hasLongRunningOperation = true;
+                    return;
                 }
+            }
+            foreach (var inputClient in rootClient.Children)
+            {
+                TraverseInput(inputClient, ref hasOperation, ref hasLongRunningOperation);
             }
         }