+++ sdk/identity/Azure.Identity/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.14.0-beta.4 (Unreleased)
+## 1.14.0-beta.5 (Unreleased)
 
 ### Features Added
 
@@ -10,7 +10,14 @@
 
 ### Other Changes
 
-- Updated `Microsoft.Identity.Client` dependency to version 4.70.2
+## 1.14.0-beta.4 (2025-05-01)
+
+### Bugs Fixed
+- Fixed an issue where Azure CLI credential could hang or delay due to I/O contention when standard input/output was shared with the host process ([#49582](https://github.com/Azure/azure-sdk-for-net/pull/49582)).
+
+### Other Changes
+
+- Updated `Microsoft.Identity.Client` dependency to version 4.71.1
 
 ## 1.14.0-beta.3 (2025-04-08)